requirements.txt
asgiref==3.3.1
azure-cognitiveservices-vision-face==0.5.0
azure-common==1.1.26
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
gcloud==0.18.3
googleapis-common-protos==1.53.0
httplib2==0.19.0
idna==2.10
isodate==0.6.0
jws==0.1.3
msrest==0.6.21
mysqlclient==2.0.3
oauth2client==4.1.3
oauthlib==3.1.0
Pillow==8.1.2
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.10.1
pyparsing==2.4.7
Pyrebase4==4.4.3
python-jwt==2.0.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
